new module dedicated to streaming tools
